Minor fix to collect_settings.js
Let's generate the broadcast notification after the user has closed
the alert, to give the FSM enough time to settle down. Otherwise, we run
into timing issues where we re-read the new value before the FSM is done
updating the state.

We should really do this in a more principled way, but don't want to embark on
a big native code rewrite at this point.

Also bump up the version number to indicate the change.

Testing done (in conjunction with
https://github.com/e-mission/e-mission-phone/commit/8deda1dbc9a38b34596c986f45d105eb83957289

- Cleared logs
- Toggled tracking
- Ensured that refreshScreen was not called
- Clicked on the button in the alert
- Ensured that refreshScreen was called and the final state was consistent with
  the toggle