Skip the settings check during `stopAll`
If we are stopping all the location tracking anyway, then do we care if there
are problems with the location?

No, we don't. Because if we do, we run into an infinite loop if we are in the
tracking_stopped state and we turn location services off. Note that this is a
perfectly valid case which should not generate an error.

But because we recheck, we generate a tracking error, which tries to get us to
stop everything, which then rechecks and generates a tracking error.... ad
infinitum.
https://github.com/e-mission/e-mission-docs/issues/567#issuecomment-681557072