Always display the current state
While experimenting with the activity permissions, the foreground service was
restarted when we got the activity permission, and we ended up with the
"background tracking" message. There is really no reason to not have message
always show the state, so let's always do that for consistency.

That also means that we can remove the newly added localized string