Add periodic checks for being stuck ongoing state without getting any points
If we are in `ongoing_trip`, but we haven't received any updates for a while,
go back to `waiting_for_trip_start`.

This is a robustness fix that allows us to deal with the case in which the
foreground service is killed for unknown reason while in the ongoing trip
state and does not resume properly.

Without this fix, we will remain stuck in the `ongoing_trip` state forever.
With this fix, we will reinitialize the FSM and be able to recover, although we
will lose the hour's worth of data during the transition.

This is a workaround/hardening solution for
https://github.com/e-mission/e-mission-docs/issues/580#issuecomment-700747931