Improve location onboarding for iOS
For iOS 13+, it is easier if the app launches the settings page to set always,
instead of the convoluted process we had earlier.

We can also check to see if the value is "Always" and prompt the user on app
launch if it is not.

+ Unify the alerting mechanisms between location and motion activity, prior to
improving the checks for motion activity.