Humanize the messages commonly generated by the foreground service
The most common messages displayed by the foreground service are the
"In state XXX" message. But the state is too jargon-y - e.g.
"In state local.state.waiting_for_trip_start".

It could also not be easily internationalized.

Based on feedback from the Smart Commute program, we changed this to more
meaningful messages. I'd been putting this off because it is a bit tricky to
do.

We have the state as a string. But the i18n solution maps from an id to the
humanized string. So how do I convert a string to an ID?
It turned out to not be that hard.

https://stackoverflow.com/a/29902630/4040267
https://stackoverflow.com/a/19093447/4040267