Fix incorrect initialization during onboarding when `markConsented` is called
This fixes
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-971953421

**Behavior before the change:**

Receive initialize as part of plugin creation

```
    public void pluginInitialize() {
        ...
        TripDiaryStateMachineReceiver.initOnUpgrade(myActivity);
    }

    public static void initOnUpgrade(Context ctxt) {
        SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(ctxt);
        System.out.println("All preferences are "+sp.getAll());

        int currentCompleteVersion = sp.getInt(SETUP_COMPLETE_KEY, 0);
        Log.d(ctxt, TAG, "Comparing installed version "+currentCompleteVersion
            + " with new version " + BuildConfig.VERSION_CODE);
        if(currentCompleteVersion != BuildConfig.VERSION_CODE) {
            Log.d(ctxt, TAG, "Setup not complete, sending initialize");
            ...
            ctxt.sendBroadcast(new ExplicitIntent(ctxt, R.string.transition_initialize));
```

```
2022-02-01 18:04:08.152 25551-25551/edu.berkeley.eecs.emission D/TripDiaryStateMachineRcvr: Setup not complete, sending initialize
```

But because the intro is not done, we exit early

```
            if (introDoneResult != null) {
                    Log.i(context, TAG, reqConsent + " is not the current consented version, skipping init...");
                    NotificationHelper.createNotification(context, STARTUP_IN_NUMBERS,
                      null, context.getString(R.string.new_data_collection_terms));
                    return;
                } else {
                Log.i(context, TAG, "onboarding is not complete, skipping prompt");
                    return;
                }
```

So far so good. We could add a check in initOnUpgrade that only generates the
initialize if the intro is not done, but that is an optimization.

Next: UI calls `markConsented`
- calls `TripDiaryStateMachineForegroundService.startProperly`
    - calls `startForegroundService` with `getForegroundServiceIntent`
        - calls `onStartCommand` which calls `humanizeState`
            ```
                    String message  = humanizeState(this, TripDiaryStateMachineService.getState(this));
            ```

The default state (since we have not initialized) is the start state, so the humanized message is `Cannot start app, see next pop up`
We don't call `initialize` after that during the onboarding process.

Let's dig a bit deeper into this.

The current data collection plugin code currently has the line `ctxt.sendBroadcast(new ExplicitIntent(ctxt, R.string.transition_initialize));`
but it was commented out as part of https://github.com/e-mission/e-mission-data-collection/commit/3e94966d4fd597bbf33273e99a421a91585d6e96
in which it was replaced by `checkAndPromptPermissions();`. The new function calls `restartFSMIfStartState` if we have the location permission.

```
    private void checkAndPromptPermissions() {
        if(cordova.hasPermission(LOCATION_PERMISSION)) {
            TripDiaryStateMachineService.restartFSMIfStartState(cordova.getActivity());
        } else {
            cordova.requestPermission(this, ENABLE_LOCATION_PERMISSION, LOCATION_PERMISSION);
        }
    }
```

and `restartFSMIfStartState` in turn generates the initialize transition.

In this refactoring, we have commented out both
`mControlDelegate.checkAndPromptPermissions` and `ctxt.sendBroadcast(new
ExplicitIntent(ctxt, R.string.transition_initialize));` so the FSM is never
initialized and the app stays in the start state.

I can confirm that if I `initialize`, the FSM starts up correctly and the message is changed.

The obvious fix is to go back to generating `initialize`. Note that there are
some checks around `intro_done` and we haven't actually finished the intro at
this point. However those checks are only triggered if the user did not
consent. Since we are in `markConsented`, those checks are not even invoked.

**New fixed flow:**
- initial `initialize` from `initOnUpgrade`, ignored since no consent and no intro_done
- `markConsented` called. This calls `setConsented` and then sends an `initialize` event
- `initalize` is received, consent is true, intro_done checks are skipped and foreground service `startProperly` is called
- foreground `startService` is called and the function returns
- foreground service `onStartCommand` is called, message is "Cannot start app, see next pop up", also shown in UI
- handle the `initialize` in the FSM
- calls `handleStart`, which sets the state to `waiting_for_trip_start`
- State Machine `setNewState` -> ForegroundServiceComm `setNewState` -> foreground service `setStateMessage`
- message is now "Ready for your next trip" both in the debugger and in the UI
- `checkAppState` is called from State Machine `setNewState`, but the app state is fine
- after onboarding, state is `waiting_for_trip_start`