Do not generate a tracking error unless the location cannot be read
We've had a mismatch about the way we deal with sensors other than location.
In the UI, we require them.
In the background checks, we generate a tracking error if they don't exist.
But in the state machine, we move to `waiting_for_trip_start` once the geofence was created successfully.

This led to an infinite loop between the `waiting_for_trip_start` and `start` states.
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1029693955

For now, we are generating a `tracking_error` only if there is no access to the
location sensor. Note that this is still inconsistent with the UI. We should
unify the UI and background options or (shudder) make it configurable.
Filed https://github.com/e-mission/e-mission-docs/issues/700 to discuss with
stakeholders and make it consistent.