Initialize the FSM after fixing permission issues
At least when the issue is fixed from the UI.

This avoids the gap between the issue being fixed, and the tracking starting again.
If the issue is fixed directly from the settings screen, we don't get a
callback, so we can't reinitialize immediately and will have to wait until the
next periodic sync.

https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1032956805
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1033118560