Initial scaffolding for the iOS status screen
Changes:
- Add the new methods to both the .m and .h files
- Create a new `NOP_RETURN_TRUE` to return TRUE from unimplemented/irrelevant methods
- Use it for `launchInit`, `fixOEMBackgroundRestrictions`, or `isNotificationsUnpaused`
- Implement `isValidShowNotifications` and `fixShowNotifications` to ensure
  that the end-to-end call works. This works as long as the user always selects
  "Allow". Still need to fix the case where the user selects "Don't Allow".

Note that the `didRegisterUserNotificationSettings` callback apparently works
from only the AppDelegate, which is problematic wrt modularity.

Will refactor the existing location and motion activity permissions first, and
then get back to notifications. May take this opportunity to move to the newer,
non-deprecated version and see if it has a smoother flow.