Fix the issue where the location option didn't show up when we opened the app page
It turns out that we need to actually start tracking the location before the
location permission would be visible, similar to the motion activity
permission
(https://github.com/e-mission/e-mission-data-collection/pull/195/commits/f3c673a9a3dda2c65802cb9656fdd751d68af8eb).

In the previous version, we would call `markConsented` and then popup
permissions as required. So the `startUpdatingLocation` call would happen in
parallel with opening up the status screen.

However, now we get all permissions upfront and then call `markConsented`. So
we need to include a dummy `startUpdatingLocation` call to see the location. We
can stop updating the location once the user has given the permission.

We didn't see this earlier (in ca979bd85f585161cdb387687bbd2acba4f3cb85)
because we tested on a physical phone which was running 12.x and can prompt for
permissions inline.

For excrutiating detail, see
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1045429690