Check the location every minute while walking instead of waiting for the next STILL transition
During initial testing, it looks like, while walking outside, the next STILL
transition can actually happen multiple minutes later. So only checking the
location at that point can lead to the trip only detected at the end.

In fact, the google geofence transition sometimes triggers before ours.
https://github.com/e-mission/e-mission-docs/issues/774#issuecomment-1223509393

This is option b) b.

>  b. re-adding the code to check the location 2 minutes after a walk
>  transition, since we will typically take 2 mins to walk 100m. This would
>  also allow us to capture the transition even if we hadn't actually stopped
>  outside the library.

Concretely:
- we schedule a location check for a minute after a walk transition, but cancel
  it if we get a STILL in the interim

This code experiments with two different options for the transition:
1) location/looper: which does not work since when the intent service is
reinitialized, it no longer has the looper variable. This would need to be
replaced with an intent service as well.
2) worker: which is scheduled to run periodically. This also needs to be a
separate public class to be initialized.

Since both need a separate class to work, we choose (2) since in that case, the
checks don't have to be location based. We can get periodic callbacks even in
the absence of location being turned on, which gives us more flexibility in our
implementation.

To avoid duplicating code, we convert the `isOutsideGeofence` call into a
static method which we can call from both the intent service and the worker.
TODO: move the checks out into their own class for better modularity

We still keep the old looper implementation around in case we want to resurrect
it, but will likely remove it before merging the code.

High level code fixes:
- `checkAlreadyOutsideGeofence` -> `isOutsideGeofence`
- convert `isOutsideGeofence` to static
    - replace all `this` with `ctxt`
- add new `scheduleDelayedCheck` which creates the action
- add new `cancelPendingDelayedCheck` which cancels the action
- create a new, unused `WalkExitGeofenceLocationCallback`