Fall back to high accuracy locations if low accuracy locations are not available
This basically implements
https://github.com/e-mission/e-mission-docs/issues/774#issuecomment-1229058627

Try to read location with balanced power accuracy
If it is not available, retry with high accuracy
In both cases, check for INSIDE vs. OUTSIDE
If even high accuracy is not available, continue to return UNKNOWN

This should fix the issue encountered in the grocery store trip
https://github.com/e-mission/e-mission-docs/issues/774#issuecomment-1226514371

I haven't been able to test this yet, the location typically seems to be known.
However, it is an optimization of an experimental feature, so let's push it out
and see how it works