Turn on op-geofence only if the user has enabled it
This is the last step of
https://github.com/e-mission/e-mission-docs/issues/774#issuecomment-1229340023

In order to keep the code footprint small for this (hopefully throwaway)
change, we assume that the user saves the enable/disable directly in local
config. This is a bad long-term assumption, but hopefully, over the long-term,
we will turn off this option and always enable it by default.

So all the changes are really in `TripDiaryStateMachineService`:
- check to see if the config exists
- if it does, turn op-geofence on/off
- it it does not, skip all op-geofence operations