:arrow_up: Set the compileSDK to 32 + bump up other library dependencies
Note that we were not able to bump up some of the dependencies to the max level
because they required a compileSDK of 33.

As seen in https://github.com/e-mission/e-mission-docs/issues/838#issuecomment-1465104696

Example:
```
Dependency 'androidx.work:work-runtime:2.8.0' requires libraries and applications that
depend on it to compile against version 33 or later of the
Android APIs.

:app is currently compiled against android-32.

Also, the maximum recommended compile SDK version for Android Gradle
plugin 7.2.1 is 32.

Recommended action: Update this project's version of the Android Gradle
plugin to one that supports 33, then update this project to use
compileSdkVerion of at least 33.
```

Testing done:
- Compiles
- Starts trip
- Need to check: trip end working?