:children_crossing: Make sure that the error message is positive
per feedback from program admin

> My point is if the text in the first screen shot that you show is under your
control to message out, don’t call it an “incorrect app setting”, call it a
“Google update, so please fix your settings!”, just something that lets them
know positively that it isn’t something wrong with the app.  Make the messaging
positive!

We don't really know that this is due to a Google update; it could also be due
to the user going into the settings at any time and changing them. But let's at
least change the messaging to be positive.

Corresponds to
https://github.com/e-mission/e-mission-phone/commit/74217affeb1bf9be5010f1dd9d40a4f506fb37ee

Not bumping up the plugin.xml since I don't want to go through the effort of
making a new release just for this change. Instead, we will incorporate it in
the next release (which should be the initial bluetooth integration).