🔥 Remove workaround to set the API version to 32
In android 10, the cordova team removed support for `compileSdkVersion`
However, in order to support auto-reset of permissions, we needed to use a
later SDK than the target.

After filing an issue in the cordova-android repo, we added this grade file as
a workaround. However, they have now reintroduced it in android 11
https://github.com/apache/cordova-android/issues/1373#issuecomment-1127157694
*and* our compile and target SDKs have caught up anyway

So we can remove the workaround.

More detail at:
https://github.com/e-mission/e-mission-phone/pull/1016#issuecomment-1685113930