🤡 Implement the testing hack as a mock method in the data collection plugin
We created a testing hack to test the previous read/write native changes.
https://github.com/e-mission/e-mission-phone/pull/1144#issuecomment-2053705589

We now move that hack to the plugin interface and make it accessible externally.
That makes it easier to call from the UI for better mocking.
Concretely, we can modify
https://github.com/e-mission/e-mission-phone/pull/1144/commits/7cb388227e52369725ab4c20c5692d433dbc6417
to call the interface, which will write sensor data using native code.
This allows us to simulate the actual working of the system (writing native
code) through the UI

Summary of changes:
- Create a new `mockBLEObjects` method in the interface
- Implement it on both android and iOS
- Change the fake object implementaton in the BluetoothBLE wrapper to take more
  arguments to that they can be passed in from the UI

Testing done:
https://github.com/e-mission/e-mission-phone/pull/1144/commits/463c2a5b642087e3901c995f8a57d4287d802067