🗃️ Save range results in the usercache
This will allow us to know *which* beacons we were seeing during the trip, and
thus, to match up the car with the trip.

Testing done: compiled

```
BUILD SUCCESSFUL in 7s
52 actionable tasks: 2 executed, 50 up-to-date
```

I am not able to test it since I don't currently have a BLE beacon, but should
be able to push it to staging so that we can test it.

Also bump up the plugin version to reflect the change