🔧 Read the fleet versus non-fleet option from the dynamic config
In https://github.com/e-mission/e-mission-data-collection/pull/220/commits/a46fad8ec52f11742f7b5631bf7dbb29ba17dd86
we hardcoded the fleet / non-fleet flag to make it easier to test
This commit reads the flag correctly from the dynamic config.

Before making this change, I was concerned that we may be calling the `init`
function before the user has put in their opcode and we have donwloaded the
dynamic config.

however, it looks, through testing, that fortunately, we have done the right thing.
we wait to initialize the FSM until the user has entered the opcode

Testing done:
https://github.com/e-mission/e-mission-data-collection/pull/220#issuecomment-2054220155

- Uninstalled and reinstalled
- TripDiaryStateMachine was not initialized
- Put in opcode
- TripDiaryStateMachine was initialized with this stacktrace

```

```

On restart (after install), the backtrace is

```
```