open app settings on notif perms prompt if initial prompt already shown
We want to give the user the option to revise their choice after their initial choice. We can only show the prompt once, so if the user triggers prompt again, we should go to app settings instead of doing nothing.
Keeps track of whether prompt already happened by storing a bool in standardUserDefaults