Single-line changes to remove notification checks in the background checker
As part of https://github.com/e-mission/e-mission-docs/issues/1094, we made the
notification permission optional.

Since it is not required for silent push notifications, notifications are
nice-to-have. They are still helpful to ensure that people know if there are
issues, and hopefully help them label their trips, but the app will still work
if the permission is turned off.

Since this is no longer required for normal operation, we don't need to check
it in the background every hour, and nag the user to fix it if it is turned off.
This is a super-easy change (two lines), so rolling this in while fixing
https://github.com/e-mission/e-mission-docs/issues/1094 instead of coming in as
a future fix.

Testing done:
- Replaced the data collection plugin with the most recent version
- Code compiles
- Wasn't able to test in the emulator since with the permission turned off, without
    https://github.com/e-mission/e-mission-phone/pull/1182 in place, we got a
    prompt to fix the notification, so we couldn't simulate a remote push

Pushing this and testing all the changes together