support deployment overrides to tracking config; add is_fleet
The tracking config controls some behaviors of this plugin. Default values are defined in the native code and differ a bit between Android and iOS. We support user-level edits but now want to support deployment-level overrides (specifically, to turn off the iOS distance filter for specific deployments https://github.com/e-mission/e-mission-docs/issues/1133)

When fleet mode was added, it was a deployment-wide setting set in the dynamic config. We handled this in ConfigManager - but separately from the other tracking config options (it got its own function "isFleet" that reads from the dynamic config)

A more future-proof, generic approach is to consider is_fleet the same as any of the other tracking options and allow *any* of them to be overridden at the deployment level

-modified ConfigManager to use deployment overrides as default values if "tracking" is preset in the deployment config
-added is_fleet to LocationTrackingConfig on both platforms with default value of false/NO, removed the isFleet function from ConfigManager and updated places where it was called
-renamed functions and variables in ConfigManager to be more aligned on both platforms (will make future work easier)
-unified try/catch behavior in ConfigManager on both platforms